requirements.txt
Flask
requests
werkzeug
python-dotenv
openai
google-cloud-texttospeech
google-cloud-speech
pydub
speechrecognition
gunicorn
